Consolidate identifier prefix test fixtures and move schema tests (#137)
